pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi",
    "pillow",
    "python-multipart",
    "ultralytics",
    "ultralyticsplus",
    "uvicorn",
    "opencv-python-headless",
]
